# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyChatbot World API
  version: 1.0.0
  description: |
    MyChatbot World 플랫폼의 전체 REST API 명세.

    ## 인증
    대부분의 엔드포인트는 Supabase 발급 JWT Bearer Token 인증이 필요합니다.
    로그인 후 수신한 `access_token`을 `Authorization: Bearer {token}` 헤더에 포함하세요.

    ## 공통 에러 형식
    ```json
    { "error": { "code": "ERROR_CODE", "message": "에러 설명", "details": {} } }
    ```
  contact:
    name: MyChatbot World Support
    url: https://mychatbot-world.vercel.app
servers:
  - url: https://mychatbot-world.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth에서 발급된 JWT 액세스 토큰
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_001
            message:
              type: string
              example: 인증이 필요합니다.
            details:
              type: object
          required:
            - code
            - message
...
paths:
...
  /api/jobs/match:
    post:
      summary: 채용 공고 지원 (POST)
      description: 프리랜서가 채용 공고에 지원합니다.
      operationId: applyJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - job_id
              properties:
                job_id:
                  type: string
                  format: uuid
                cover_letter:
                  type: string
                proposed_amount:
                  type: integer
      responses:
        '201':
          description: 지원 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  match_id:
                    type: string
                    format: uuid
        '409':
          description: 이미 채용 완료된 공고
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...